% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Create a data.frame containing the raw NOAA Significant Earthquakes Dataset}
\usage{
eq_clean_data(earthquakes_raw)
}
\arguments{
\item{earthquakes_raw}{The raw NOAA data frame to be cleaned.}
}
\value{
This function returns a data.frame containing the cleaned NOAA earthquake data.
A DATE column is created by uniting the year, month day and converting it to Date class.
LATITUDE and LONGITUDE columns are converted to numeric class and
the LOCATION_NAME column has been cleaned by stripping out the country name (including the colon)
and converting the names to title case (as opposed to all caps).
}
\description{
This function reads in the earthquake data from the U.S. National Oceanographic
and Atmospheric Administration (NOAA) on significant earthquakes around the world.
This dataset contains information about 5,933 earthquakes over an approximately 4,000 year time span.
and is avaialable at the URL:
https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
}
\details{
This function takes the raw NOAA data frame and returns a clean data frame.
The clean data frame has the following:

1.	A date column created by uniting the year, month, day and converting it to the Date class
2.	LATITUDE and LONGITUDE columns converted to numeric class
3.	The LOCATION_NAME column is cleaned by stripping out the country name (including the colon)
   and converting the names to title case (as opposed to all caps).
}
\examples{
\dontrun{eq_clean_data(earthquakes.tsv)}

}
