% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{eq_map}
\alias{eq_map}
\title{Maps the epicenter (LATITUDE/LONGITUDE) of each earthquake in a data frame and annotates each point with a pop up window.}
\usage{
eq_map(data = NULL, annot_col = "DATE")
}
\arguments{
\item{data}{The data to be used to create the points on the map.

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot.

A data.frame, or other object, will override the plot data.  All objects will be fortified to produce a data frame.  See fortify for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.}

\item{annot_col}{A text string containing the name of the column in the data frame to use for the annotation text.}
}
\value{
This function has no return value
}
\description{
This function takes an argument, data, containing a data frame with earthquakes to visualize.
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates each point with in pop up window
containing annotation data stored in a column of the data frame. The user may choose which column is used
for the annotation in the pop-up with a function argument named annot_col.
Each earthquake is shown with a circle, and the radius of the circle is proportional to the earthquake's magnitude (EQ_PRIMARY)
}
\examples{
\dontrun{eq_map(data = eq_data, annot_col = "DATE")}

}
