% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Geom for adding annotations to earthquake data displyed with geom_timeline}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, parse = FALSE, nudge_x = 0, nudge_y = 20,
  check_overlap = FALSE, x = NULL, label = NULL, n_max = NA,
  xmindate = NA, xmaxdate = NA, ...)
}
\arguments{
\item{mapping}{A set of aesthetic mappings created by aes or aes_.  If specified and inherit.aes = TRUE (the default),
it is combined with the default mappint at the top level of the plot.  You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.  There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot.

A data.frame, or other object, will override the plot data.  All objects will be fortified to produce a data frame.
See fortify for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are mremoved with a warning.  If TRUE, missing values are silently removed.}

\item{show.legend}{logical.  Should this layer be included in the legends?  NA, the default, incldes if any aesthetics are mapped.
False never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.
This is most usefule for helper functions that
define both data and aesthetics and shouldn't inherit behavior from the default plot specification, e.t. borders.}

\item{parse}{If TRUE, the labels will be parsed into expressions and displayed
as described in ?plotmath}

\item{nudge_x}{Horizontal offset for text from point}

\item{nudge_y}{Vertical offset for text from point}

\item{check_overlap}{If TRUE, text that overlaps previous text in the same layer will not be plotted}

\item{x}{Date of the earthquake (required)}

\item{label}{Name of the column to be used for the annotation text.}

\item{n_max}{Number of largest earthquakes (by magnitude) that
are included in the subset of earthquakes that are annotated.}

\item{xmindate}{Minimum year to display on the timeline}

\item{xmaxdate}{Maximum year to display on the timeline}

\item{...}{Other arguments passed on layer.  Thes are often aesthetics, used to set an aesthetic to a fixed value,
like color = "red" or size = 3.  They may also be parameters to the paired geom/stat.}
}
\value{
This function has no return value
}
\description{
This function creates a new geom which will add annotations to the earthquake data displayed by
the geom_timeline geom.  This geom adds a vertical line to each data point with a text annotation
(e.g. the location of the earthquake) attached to each line.  There is an option to subset
the data to n_max number of earthquakes, where only the n_max number of largest earthquakes (by magnitude)
are included in the subset.  Aesthetics include x, which is the date of the earthquake, and label, which
takes the column name from which annotations are obtained.
}
\details{
This code is based on input from the Extending ggplot2 vignette:
https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html
}
\examples{
\dontrun{geom_timeline_label(data = earthquakes_clean,
                       aes(x = DATE, label = LOCATION_NAME)}

}
