% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\docType{data}
\name{GeomEarthquake}
\alias{GeomEarthquake}
\title{The following code creates the geom_timeline geom that is used to visualize some of the
information in the NOAA earthquakes dataset.}
\format{An object of class \code{GeomEarthquake} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomEarthquake
}
\description{
Create a new ggproto object (GeomEarthquake) for a new geom (geom_timeline) for ggplot2
}
\details{
The following code creates a new geom which will add a hurricane wind radii chart
to a base map.  The user must supply the desired longitude and latitude for the
wind radii chart as well as wind radius information for three wind speeds
(34 knots, 50 knots and 64 knots) each of the four compampass
quadrants (ne, nw, se, sw)

This ggproto object inherits from the basic Geom.  Default values are provided for the
aesthetics size, linetype and alpha.  A null ggproto opject is created if there are not
three or more points in the polygon.
}
\keyword{datasets}
