% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_functions.R
\name{stat_timeline_label}
\alias{stat_timeline_label}
\title{Stat for creating a timeline for earthquake data labels}
\usage{
stat_timeline_label(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, x = NULL, label = NULL, xmindate = NULL,
  xmaxdate = NULL, n_max = NA, ...)
}
\arguments{
\item{mapping}{A set of aesthetic mappings created by aes or aes_.  If specified and inherit.aes = TRUE (the default),
it is combined with the default mappint at the top level of the plot.  You must supply mapping if there is no plot mapping.}

\item{data}{The data to be displayed in this layer.  There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot.

A data.frame, or other object, will override the plot data.  All objects will be fortified to produce a data frame.
See fortify for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data.}

\item{geom}{The geom to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are mremoved with a warning.  If TRUE, missing values are silently removed.}

\item{show.legend}{logical.  Should this layer be included in the legends?  NA, the default, incldes if any aesthetics are mapped.  False never includes,
and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.  This is most usefule for helper functions that
define both data and aesthetics and shouldn't inherit behavior from the default plot specification, e.t. borders.}

\item{x}{Date of the earthquake (required)}

\item{label}{Name of column to use for labels (required)
#@param y Factor indicating some striatification in which case multiple time lines will be ploted for each level of the factor (e.g. country).}

\item{xmindate}{Minimum year to display on the timeline}

\item{xmaxdate}{Maximum year to display on the timeline}

\item{n_max}{The maximum number of (largest) earthquakes to include on the timeline}

\item{...}{Other arguments passed on layer.  Thes are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3.
They may also be parameters to the paired geom/stat.}
}
\value{
This function has no return value
}
\description{
This function creates a new stat which will create a linear timeline for a
specified time range, and will label each earthquake (up to n_max) as a label
connected with a line to a point on the timeline.
}
\details{
This code is based on input from the Extending ggplot2 vignette:
https://cran.r-project.org/web/packages/ggplot2/vignettes/extending-ggplot2.html
}
\examples{
\dontrun{stat_timeline_label(data = earthquakes_clean,
                       aes(x = DATE, label = LOCATION_NAME),
                       xmindate = 2000, xmaxdate = 2017, n_max = 5)}

}
\keyword{internal}
